+++ allensdk/brain_observatory/behavior/session_apis/data_io/behavior_json_api.py
@@ -51,23 +51,19 @@ class BehaviorJsonExtractor(BehaviorDataExtractorBase):
         return self.data['age']
 
     def get_stimulus_name(self) -> str:
-        """Get the name of the stimulus presented for a behavior or
-        behavior + ophys experiment"""
+        """Get the name of the stimulus presented for the experiment"""
         return self.data['stimulus_name']
 
     def get_reporter_line(self) -> str:
-        """Get the (gene) reporter line for the subject associated with a
-        behavior or behavior + ophys experiment"""
+        """Get the (gene) reporter line for the subject associated with an experiment"""
         return self.data['reporter_line']
 
     def get_driver_line(self) -> str:
-        """Get the (gene) driver line for the subject associated with a
-        behavior or behavior + ophys experiment"""
+        """Get the (gene) driver line for the subject associated with an experiment"""
         return self.data['driver_line']
 
     def get_full_genotype(self) -> str:
-        """Get the full genotype of the subject associated with a
-        behavior or behavior + ophys experiment"""
+        """Get the full genotype of the subject associated with an experiment"""
         return self.data['full_genotype']
 
     def get_behavior_stimulus_file(self) -> str:
@@ -78,3 +74,9 @@ class BehaviorJsonExtractor(BehaviorDataExtractorBase):
         """Get the external specimen id (LabTracks ID) for the subject
         associated with a behavior experiment"""
         return int(self.data['external_specimen_name'])
+
+    def get_experiment_date(self) -> datetime:
+        """Get the acquisition date of an experiment"""
+        return pytz.utc.localize(
+            datetime.strptime(self.data['date_of_acquisition'],
+                              "%Y-%m-%d %H:%M:%S"))